+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Implement deleting clusters and kapps  
 * Change the args field from a list of maps to just maps
 * Added a 'kapps clean' subcommand to run 'make clean' across selected kapps
+* Added a 'kapps output' subcommand to run 'make output' across selected kapps
 
 ## 0.6.0 (25/3/19)
 * Major code clean up & refactoring